AppWeb/requirements.txt
# requirements.txt
flask
flask-login
pyodbc
bcrypt
datetime
plotly
re
io
azure-storage-blob
